Refactor websocket handlers into modules
